refactor: better file hierarchy
