+++ src/util/mod.rs
@@ -1,0 +1,10 @@
+//! Utilities
+
+pub mod adjust_pwd;
+pub mod echo;
+pub mod encode;
+pub mod file_type;
+pub mod r#match;
+pub mod request;
+pub mod response;
+pub mod result;
