+++ src/util/result.rs
